feature/KLDV-3068 KlipDrops API 호출: apply review

+++ src/main/java/com/klipwallet/membership/adaptor/klipdrops/dto/Drops.java
@@ -6,6 +6,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record Drops(
         @JsonProperty("drops") List<Drop> drops,
-        @JsonProperty("totalCount") Integer totalCount
+        @JsonProperty("totalCount") Long totalCount
 ) {
 }
